delete post and category

+++ frontend/src/pages/DeletePost.jsx
@@ -11,22 +11,19 @@ function DeletePost() {
   const access_token = user?.data?.token;
   const { postId } = useParams();
   const navigate = useNavigate();
-  console.log("Categoria id", post?.category);
-
   const categoryId = post?.category;
-  console.log("Category id", categoryId);
+  
   useEffect(() => {
     const fetchData = async () => {
       try {
-        // Ellenőrizzük, van-e access_token
-        if (access_token) {
+        // Ellenőrizzük, van-e access_token és categoryId
+        if (access_token && categoryId) {
           // Lekérjük a kategória részleteit a categoryId alapján
           const categoryDetails = await CategoryService.singleCategory(
             access_token,
             categoryId
           );
           setCategory(categoryDetails);
-          console.log("Kategória részletei:", categoryDetails);
         }
       } catch (error) {
         console.error("Hiba a kategória részleteinek lekérésekor:", error);
@@ -34,23 +31,7 @@ function DeletePost() {
     };
     fetchData();
   }, [access_token, categoryId]);
-
-  // categoria törlése
-  const handleDeleteCategory = async () => {
-    try {
-      // Ellenőrizzük, van-e access_token, a post elérhető-e, és az ID-k egyeznek-e
-      if (access_token && category && category._id === categoryId) {
-        // Hívjuk meg a service metódust a poszt törléséhez
-        await CategoryService.deleteCategory(access_token, categoryId);
-        // Sikeres törlés után átirányítjuk a felhasználót a "/posts" útvonalra
-        navigate("/");
-      }
-    } catch (error) {
-      console.error("Hiba a poszt törlésekor:", error);
-    }
-  };
-
-  // Log mindkét azonosítót a konzolra
+  
   useEffect(() => {
     const fetchData = async () => {
       try {
@@ -62,13 +43,6 @@ function DeletePost() {
             postId
           );
           setPost(postDetails);
-
-          // Ellenőrizzük, hogy a postDetails objektumban található id egyezik-e a postId-vel
-          if (postDetails && postDetails.id === postId) {
-            // console.log("A postId megegyezik a postDetails objektum id-jével!");
-            //console.log("postId:", postId);
-            // console.log("postDetails.id:", postDetails.id);
-          }
         }
       } catch (error) {
         console.error("Hiba a poszt részleteinek lekérésekor:", error);
@@ -77,25 +51,28 @@ function DeletePost() {
     fetchData();
   }, [access_token, postId]);
 
-  {
-    /** 
-  // Poszt törlése
-  const handleDelete = async () => {
+  // categoria törlése
+  const handleDeletePost = async () => {
     try {
-      // Ellenőrizzük, van-e access_token, a post elérhető-e, és az ID-k egyeznek-e
-      if (access_token && post && post.id === postId) {
-        // Hívjuk meg a service metódust a poszt törléséhez
+      // Ellenőrizzük, van-e access_token, a post és a kategória elérhető-e, és az ID-k egyeznek-e
+      if (
+        access_token &&
+        post &&
+        category &&
+        post.id === postId &&
+        category._id === categoryId
+      ) {
+        // Hívjuk meg a service metódust a kategória törléséhez
+        await CategoryService.deleteCategory(access_token, categoryId);
+        // Hívjuk meg a service metódust a post törléséhez
         await PostService.deletePost(access_token, postId);
         // Sikeres törlés után átirányítjuk a felhasználót a "/posts" útvonalra
         navigate("/");
       }
     } catch (error) {
-      console.error("Hiba a poszt törlésekor:", error);
+      console.error("Hiba a törlés közben:", error);
     }
   };
-
-  */
-  }
 
   return (
     <div>
@@ -106,7 +83,7 @@ function DeletePost() {
           <p>Tartalom: {post.content}</p>
         </div>
       )}
-      <button onClick={handleDeleteCategory}>Törlés</button>
+      <button onClick={handleDeletePost}>Törlés</button>
     </div>
   );
 }
